Struct parity_hash::H2048

#[repr(C)]

pub struct H2048(pub [u8; 256]);

Unformatted binary data of fixed length.

Methods

impl H2048

pub fn new() -> H2048

Create a new, zero-initialised, instance.

pub fn zero() -> H2048

Synonym for new(). Prefer to new as it's more readable.

pub fn len() -> usize

Get the size of this object in bytes.

pub fn as_ptr(&self) -> *const u8

Returns a constant raw pointer to the value

pub fn as_mut_ptr(&mut self) -> *mut u8

Returns a mutable raw pointer to the value

pub fn clone_from_slice(&mut self, src: &[u8]) -> usize

Assign self to be of the same value as a slice of bytes of length len().

pub fn from_slice(src: &[u8]) -> Self

Convert a slice of bytes of length len() to an instance of this type.

pub fn copy_to(&self, dest: &mut [u8])

Copy the data of this object into some mutable slice of length len().

pub fn contains<'a>(&'a self, b: &'a Self) -> bool

Returns true if all bits set in b are also set in self.

pub fn is_zero(&self) -> bool

Returns true if no bits are set.

pub fn low_u64(&self) -> u64

Returns the lowest 8 bytes interpreted as a BigEndian integer.